I have no idea what im doing

+++ src/main/java/sw_10/p3_backend/Logic/ScheduleLogic.java
@@ -49,14 +49,19 @@ public class ScheduleLogic {
 
 
         //clone active schedule
-        Schedule newViewSchedule = (Schedule) CurrentEditschedule.clone();
+        Schedule newViewSchedule = (Schedule) CurrentEditschedule.clone(true);
 
         //set the new schedule to viewable (not active)
         newViewSchedule.setActive(true);
 
         //TODO: Not secure. If something goes wrong is the old schedule lost and the new one is not saved.
         //save the new schedule (will be the new view only schedule) and delete the old one.
-        scheduleRepository.delete(CurrentViewSchedule);
+        System.out.println(CurrentViewSchedule);
+        if(CurrentViewSchedule != null) {
+            scheduleRepository.delete(CurrentViewSchedule);
+            System.out.println("Deleted");
+        }
+
         scheduleRepository.save(newViewSchedule);
 
         return newViewSchedule;
@@ -68,5 +73,27 @@ public class ScheduleLogic {
         System.out.println("deleteSchedule");
         scheduleRepository.deleteById(Long.valueOf(id));
         return schedule;
+    }
+
+    public Schedule discardEditChanges()  throws CloneNotSupportedException {
+
+        //Find editable schedule (active)
+        Schedule CurrentEditschedule = scheduleRepository.findScheduleByIsActive(false);
+        Schedule CurrentViewSchedule = scheduleRepository.findScheduleByIsActive(true);
+
+        //delete editable schedule
+        scheduleRepository.delete(CurrentEditschedule);
+
+        //set the viewable schedule to editable
+        Schedule newEditSchedule = (Schedule) CurrentViewSchedule.clone(false);
+
+        newEditSchedule.setActive(false);
+
+        //save the new schedule (will be the new view only schedule) and delete the old one.
+        scheduleRepository.save(newEditSchedule);
+
+        cloneScheduleAndReplace();
+
+        return newEditSchedule;
     }
 }
